Move Callable Pytypes to Directory
