+++ pystd/src/builtin_runtime.cpp
@@ -9,6 +9,7 @@
 #include <iostream>
 #include <ranges>
 
+#include "pytypes/callables/py_function.hpp"
 #include "pytypes/iterables/py_dict.hpp"
 #include "pytypes/iterables/py_list.hpp"
 #include "pytypes/iterables/py_set.hpp"
@@ -18,7 +19,6 @@
 #include "pytypes/primitives/py_int.hpp"
 #include "pytypes/primitives/py_none.hpp"
 #include "pytypes/primitives/py_str.hpp"
-#include "pytypes/py_function.hpp"
 #include "runtime_errors.hpp"
 #include "runtime_state.hpp"
 #include "runtime_util.hpp"
