+++ pystd/src/pytypes/iterables/py_set.cpp
@@ -6,12 +6,12 @@
 
 #include <format>
 
+#include "pytypes/callables/py_method.hpp"
 #include "pytypes/iterables/py_list.hpp"
 #include "pytypes/iterables/py_tuple.hpp"
 #include "pytypes/primitives/py_bool.hpp"
 #include "pytypes/primitives/py_int.hpp"
 #include "pytypes/primitives/py_none.hpp"
-#include "pytypes/py_method.hpp"
 #include "runtime_errors.hpp"
 #include "runtime_util.hpp"
